+++ iscc_schema/schema.py
@@ -557,7 +557,7 @@ class IsccJsonld(BaseModel):
     """
 
     context_: AnyUrl | None = Field(
-        "http://purl.org/iscc/context",
+        "http://purl.org/iscc/context/0.5.0.jsonld",
         alias="@context",
         description="The [JSON-LD](https://json-ld.org/) Context URI for ISCC metadata.",
     )
@@ -567,7 +567,7 @@ class IsccJsonld(BaseModel):
         description="The type of digital content according to schema.org classes (TextDigitalDocument, ImageObject, AudioObject, VideoObject).",
     )
     schema_: AnyUrl | None = Field(
-        "http://purl.org/iscc/schema",
+        "http://purl.org/iscc/schema/0.5.0.json",
         alias="$schema",
         description="The [JSON Schema](https://json-schema.org/) URI for ISCC metadata.",
         json_schema_extra={"x-iscc-context": "http://purl.org/iscc/terms/#$schema"},