+++ iscc_schema/service_tdm.py
@@ -61,8 +61,8 @@ class TDM(BaseModel):
     model_config = ConfigDict(
         extra="forbid",
     )
-    context_: Literal["http://purl.org/iscc/context"] = Field(
-        "http://purl.org/iscc/context",
+    context_: Literal["http://purl.org/iscc/context/0.5.0.jsonld"] = Field(
+        "http://purl.org/iscc/context/0.5.0.jsonld",
         alias="@context",
         description="The JSON-LD Context URI for ISCC metadata.",
     )