+++ tests/test_iscc_meta.py
@@ -1,5 +1,10 @@
 # -*- coding: utf-8 -*-
+import iscc_schema
 from iscc_schema import IsccMeta
+
+V = iscc_schema.__version__
+SCHEMA_URL = f"http://purl.org/iscc/schema/{V}.json"
+CONTEXT_URL = f"http://purl.org/iscc/context/{V}.jsonld"
 
 
 def test_iscc_meta_empty():
@@ -7,10 +12,12 @@ def test_iscc_meta_empty():
     assert m.iscc is None
     assert m.dict() == {}
     assert (
-        m.json() == '{"@context":"http://purl.org/iscc/context","@type":"CreativeWork",'
-        '"$schema":"http://purl.org/iscc/schema"}'
+        m.json()
+        == f'{{"@context":"{CONTEXT_URL}","@type":"CreativeWork","$schema":"{SCHEMA_URL}"}}'
     )
     assert (
-        m.jcs() == b'{"$schema":"http://purl.org/iscc/schema","@context":"http://purl.'
-        b'org/iscc/context","@type":"CreativeWork"}'
+        m.jcs()
+        == f'{{"$schema":"{SCHEMA_URL}","@context":"{CONTEXT_URL}","@type":"CreativeWork"}}'.encode(
+            "utf-8"
+        )
     )